requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0
# psycopg2-binary==2.9.9  # Removido temporariamente para desenvolvimento local
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
